kernel review
