+++ kernel/proto/plan.proto
@@ -34,6 +34,9 @@ message ScanParquetNode {
   repeated ScanFile files = 1;
   repeated string file_constant_columns = 2;
   delta.kernel.schema.StructType schema = 3;
+  // True preserves input-file and per-file row order, with each batch containing rows from at
+  // most one file. False, including the proto3 default, provides no order or file-boundary
+  // guarantee.
   bool ordered_scan = 4;
 }
 
@@ -41,6 +44,9 @@ message ScanJsonNode {
   repeated ScanFile files = 1;
   repeated string file_constant_columns = 2;
   delta.kernel.schema.StructType schema = 3;
+  // True preserves input-file and per-file row order, with each batch containing rows from at
+  // most one file. False, including the proto3 default, provides no order or file-boundary
+  // guarantee.
   bool ordered_scan = 4;
 }
