+++ kernel/src/plans/proto/convert.rs
@@ -1366,34 +1366,38 @@ mod tests {
         assert_eq!(proto.file_constants.len(), 2);
     }
 
-    #[test]
-    fn from_scan_parquet() {
+    #[rstest]
+    #[case(false)]
+    #[case(true)]
+    fn from_scan_parquet(#[case] ordered_scan: bool) {
         let node = ScanParquet {
             files: vec![ScanFile::new(sample_file_meta())],
             file_constant_columns: vec!["c".to_string()],
             schema: sample_schema(),
-            ordered_scan: true,
+            ordered_scan,
         };
         let proto = proto_plan::ScanParquetNode::from(&node);
         assert_eq!(proto.files.len(), 1);
         assert_eq!(proto.file_constant_columns.len(), 1);
         assert!(proto.schema.is_some());
-        assert!(proto.ordered_scan);
+        assert_eq!(proto.ordered_scan, ordered_scan);
     }
 
-    #[test]
-    fn from_scan_json() {
+    #[rstest]
+    #[case(false)]
+    #[case(true)]
+    fn from_scan_json(#[case] ordered_scan: bool) {
         let node = ScanJson {
             files: vec![ScanFile::new(sample_file_meta())],
             file_constant_columns: vec!["c".to_string()],
             schema: sample_schema(),
-            ordered_scan: true,
+            ordered_scan,
         };
         let proto = proto_plan::ScanJsonNode::from(&node);
         assert_eq!(proto.files.len(), 1);
         assert_eq!(proto.file_constant_columns.len(), 1);
         assert!(proto.schema.is_some());
-        assert!(proto.ordered_scan);
+        assert_eq!(proto.ordered_scan, ordered_scan);
     }
 
     #[test]
